requirements.txt
absl-py==1.2.0
aniso8601==9.0.1
astunparse==1.6.3
attrs==22.1.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
flask-restplus==0.13.0
flatbuffers==1.12
gast==0.4.0
gevent==21.12.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.3
grpcio==1.47.0
gunicorn==20.1.0
h5py==3.7.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.14.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.22.2
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.2
termcolor==1.1.0
torch==1.12.1
torchvision==0.13.1
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.4.0